--- src/apps/mean-finance/helpers/tokens.ts
@@ -1,15 +1,0 @@
-export const sortTokens = (tokenA: string, tokenB: string) => {
-  let token0 = tokenA;
-  let token1 = tokenB;
-
-  if (tokenA > tokenB) {
-    token0 = tokenB;
-    token1 = tokenA;
-  }
-
-  return [token0, token1];
-};
-
-export const TOKEN_TYPE_BASE = 'BASE';
-export const TOKEN_TYPE_WRAPPED = 'WRAPPED_PROTOCOL_TOKEN';
-export const TOKEN_TYPE_YIELD_BEARING_SHARES = 'YIELD_BEARING_SHARES';
