+++ src/apps/mean-finance/mean-finance.definition.ts
@@ -9,6 +9,8 @@ export const MEAN_FINANCE_DEFINITION = appDefinition({
   description:
     'Mean Finance enables users to Dollar Cost Average (DCA) any ERC20 into any ERC20 with their preferred period frequency.',
   url: 'https://mean.finance',
+  tags: [AppTag.ASSET_MANAGEMENT, AppTag.CROSS_CHAIN, AppTag.PREDICTION_MARKET],
+  keywords: [],
 
   groups: {
     dcaPosition: {
@@ -18,8 +20,6 @@ export const MEAN_FINANCE_DEFINITION = appDefinition({
     },
   },
 
-  tags: [AppTag.ASSET_MANAGEMENT, AppTag.CROSS_CHAIN, AppTag.PREDICTION_MARKET],
-  keywords: [],
   links: {
     discord: 'https://discord.com/invite/ThfzDdn4pn/',
     github: 'https://github.com/Mean-Finance',
@@ -41,5 +41,3 @@ export class MeanFinanceAppDefinition extends AppDefinition {
     super(MEAN_FINANCE_DEFINITION);
   }
 }
-
-export default MEAN_FINANCE_DEFINITION;
