+++ src/apps/mean-finance/mean-finance.module.ts
@@ -1,13 +1,10 @@
 import { Register } from '~app-toolkit/decorators';
 import { AbstractApp } from '~app/app.dynamic-module';
 
-import { ArbitrumMeanFinanceBalanceFetcher } from './arbitrum/mean-finance.balance-fetcher';
 import { ArbitrumMeanFinanceDcaPositionContractPositionFetcher } from './arbitrum/mean-finance.dca-position.contract-position-fetcher';
 import { MeanFinanceContractFactory } from './contracts';
 import { MeanFinanceAppDefinition, MEAN_FINANCE_DEFINITION } from './mean-finance.definition';
-import { OptimismMeanFinanceBalanceFetcher } from './optimism/mean-finance.balance-fetcher';
 import { OptimismMeanFinanceDcaPositionContractPositionFetcher } from './optimism/mean-finance.dca-position.contract-position-fetcher';
-import { PolygonMeanFinanceBalanceFetcher } from './polygon/mean-finance.balance-fetcher';
 import { PolygonMeanFinanceDcaPositionContractPositionFetcher } from './polygon/mean-finance.dca-position.contract-position-fetcher';
 
 @Register.AppModule({
@@ -15,11 +12,8 @@ import { PolygonMeanFinanceDcaPositionContractPositionFetcher } from './polygon/
   providers: [
     MeanFinanceAppDefinition,
     MeanFinanceContractFactory,
-    OptimismMeanFinanceBalanceFetcher,
     OptimismMeanFinanceDcaPositionContractPositionFetcher,
-    PolygonMeanFinanceBalanceFetcher,
     PolygonMeanFinanceDcaPositionContractPositionFetcher,
-    ArbitrumMeanFinanceBalanceFetcher,
     ArbitrumMeanFinanceDcaPositionContractPositionFetcher,
   ],
 })
